Bugfixes & Root Redirect Helper
This fixes some bugs in the admin with creating and deleting variables.
Unfortunately the delete button was inside the form causing issues.

This also adds a helper to assist in redirecting the root of the app
in case you want to serve a different page on a subdomain as the home.